refactored ChessPiece to not take position in constructor
This allows us to create a ChessPiece without having to assign it a position. It also solves pattern matching issues where
positions are considered as part of the pattern match when all we care for is the Piece and not its content. This can be a hindrance
when trying to get distinct collections where 2 similar pieces in the same slot in different configurations e.g. switching 2 Knight on the board should
not make a different configuration. Changed all tests associated with constructors for the ChessPieces