added a placePieceLooper function and some tests for it,the function uses backtracking to place a list of pieces on the board
It takes a list of pieces to place List[ChessPiece], a function to check if a piece can safely be placed on a slot-
the function is reseeded with with new data every time a new piece is place and passed through the arguments during the recursion,
an index of type Int to state where in the chessboard the function will start placing pieces,a vector of Positions we can't go due
to other pieces having them under attack - this placesWeCantGo:vector[Position] is updated every time we do a recursion to accomodate
slots made unsafe by adding the last piece, and a placePieces list which we use to carry state of the pieces that have already been placed
on the chessboard - we append to it when we succeed in placing a piece and pop it when we are backtracking. If we succeeded, the items in
placedPieces match the ones in the input list of List[ChessPieces] though in a reverse order.