Enabled creation of different permutations and configuration of input chesspieces
Added a method createPermutationsOfInput which takes a list of chesspieces as input and outputs different permutations  in a list.
For each of the permutations we'll try to lay out different configurations.
Also added shiftStartIndexVariations() which works by varying the starting point of where we lay the pieces on the board. This way we create
different configuration by starting at incrementing point on the chessboard i.e. for every configuration we try laying out the pieces by
starting at index 0, then start again at index 1 till we get to the end of the board.